# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Books",
        "description": "Rotas da API de Livros.",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://apibooks.mauridesenvolvedor.com/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/orders": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "ORDERS",
                "description": "Rota para trazer e filtrar os pedidos",
                "operationId": "f2ed02d05432496b7f9754441e208e66",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Buscar por data (Formato Y-m-d)",
                        "required": false
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Buscar por status (pendent,complete,cancel)",
                        "required": false
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "REGISTRO DO PEDIDO",
                "description": "Rota para registro do pedido",
                "operationId": "faf8f1ab3266846b03c01ffae9b09147",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "OrderRequest": {
                "title": "Auth Register Request",
                "description": "order request body data",
                "required": [
                    "total",
                    "orders_items"
                ],
                "properties": {
                    "total": {
                        "title": "total",
                        "description": "Total do pedido",
                        "type": "number",
                        "format": "float",
                        "example": "21.98"
                    },
                    "orders_items": {
                        "title": "order_items",
                        "description": "Items do pedido",
                        "type": "array",
                        "items": {
                            "required": [
                                "id_book",
                                "quantity",
                                "unit_price",
                                "total_price"
                            ],
                            "properties": {
                                "id_book": {
                                    "description": "ID do livro",
                                    "type": "integer",
                                    "example": 1
                                },
                                "quantity": {
                                    "description": "Quantidade do livro",
                                    "type": "integer",
                                    "example": 2
                                },
                                "unit_price": {
                                    "description": "Preço unitário do livro",
                                    "type": "number",
                                    "format": "float",
                                    "example": 10.99
                                },
                                "total_price": {
                                    "description": "Preço total do livro",
                                    "type": "number",
                                    "format": "float",
                                    "example": 21.98
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}